refactor: access WysiwygProfile
